Foram implementadas as Queries no Projeto
- Foram deletadas projeções que não seriam mais usadas;
- Alterou-se as queries que estavam em outras entidades para o projeto;
- A implementação foi feita nas demais camadas.

+++ src/main/java/br/senac/bluhelp/dto/project/ProjectProjectionDTO.java
@@ -7,5 +7,5 @@ import br.senac.bluhelp.enumeration.progress.Progress;
 import br.senac.bluhelp.projection.category.CategoryProjection;
 
 public record ProjectProjectionDTO(Long id, Long creator, String creatorName, String creatorSurname, String title, 
-		String objective, Long address, String district, String description, List<CategoryProjection> categories, Progress progress, byte[] photo, LocalDateTime date, float averageReview) {
+		String objective, Long address, String district, String description, List<CategoryProjection> categories, Progress progress, byte[] photo, LocalDateTime date, Double averageReview) {
 }
